Immediately display trigger content
Before this patch, trigger content was displayed once remote settings
and post data were fetched, thus creating a blinking effect.

This patch no longer wait for remote settings or post data in order to
display the trigger content. As a consequence, when a user clicks on the
trigger, not all data are necessarily available yet. In that case
a spinner is displayed.

Two CSS custom properties have been introduced to control the color of
the spinner.